Add logger

+++ rusty/krusty/src/kernelib.rs
@@ -10,13 +10,13 @@ extern "C" {
     pub fn kalloc() -> *mut u8;
     pub fn kfree(_: *mut u8);
     pub fn argraw(_: uint) -> usize;
-    pub fn argstr(_: uint, _: *mut u8, max: uint) -> isize;
+    pub fn argstr(_: uint, _: *mut u8, max: uint) -> int;
     pub fn get_pagetable() -> pagetable_t;
     pub fn copyout(_: pagetable_t, dst: *mut u8, src: *const u8, len: usize) -> int;
     pub fn mappages(pagetable: pagetable_t, va: usize, size: usize, pa: usize, perm: int) -> int;
 }
 
-pub fn argstr_sys(n: usize, buf: &mut [u8]) -> isize {
+pub fn argstr_sys(n: usize, buf: &mut [u8]) -> int {
     unsafe {
         argstr(n as _, buf.as_mut_ptr(), buf.len() as _)
     }
